Added support for numbers in expressions, for example, "%4n" matches "1234"

+++ match.h
@@ -88,7 +88,7 @@ struct Match_Char {
 	enum Match_Type type;
 	bool isUpper;
 	char c;
-	size_t n;
+	long n;
 };
 
 struct Match_State {
